fix: v3.15.1 switch 字段类型 int 化（防 PG boolean 分叉）
- CrudDesigner switch type tinyint->int：引擎 default=1 的 tinyint 落 PG boolean，
  与 0/1 字典/渲染分叉（列表状态列空白）；int 后迁移与引擎 DDL 同构

+++ src/support/CrudDesigner.php
@@ -245,7 +245,8 @@ class CrudDesigner
                 $field['type'] = 'text';
                 break;
             case 'switch':
-                $field['type'] = 'tinyint';
+                // int 0/1（勿用 tinyint：radmin 引擎 default='1'+tinyint 会落 PG boolean，与迁移/字典 0/1 分叉）
+                $field['type'] = 'int';
                 $field['unsigned'] = true;
                 $field['null'] = false;
                 $field['defaultType'] = 'INPUT';
@@ -537,7 +538,7 @@ class CrudDesigner
                 $opts += ['null' => true];
                 break;
             case $f['type'] === 'tinyint':
-                // PG 无 tinyint：integer 兜底（switch 布尔语义用 0/1）
+                // 兼容历史（switch 已统一 int；tinyint 落 integer）
                 $type = 'integer';
                 $opts += ['signed' => !$f['unsigned'], 'null' => false];
                 $opts['default'] = $f['defaultType'] === 'INPUT' ? (int) ($f['default'] === '' ? '0' : $f['default']) : 0;
